+++ core/src/main/java/cucumber/runtime/table/TableDiffer.java
@@ -14,13 +14,13 @@ import java.util.Map;
 
 public class TableDiffer {
 
-    private final DataTable orig;
-    private final DataTable other;
+    private final DataTable from;
+    private final DataTable to;
 
-    public TableDiffer(DataTable origTable, DataTable otherTable) {
-        checkColumns(origTable, otherTable);
-        this.orig = origTable;
-        this.other = otherTable;
+    public TableDiffer(DataTable fromTable, DataTable toTable) {
+        checkColumns(fromTable, toTable);
+        this.from = fromTable;
+        this.to = toTable;
     }
 
     private void checkColumns(DataTable a, DataTable b) {
@@ -30,11 +30,11 @@ public class TableDiffer {
     }
 
     public void calculateDiffs() throws TableDiffException {
-        Patch patch = DiffUtils.diff(orig.diffableRows(), other.diffableRows());
+        Patch patch = DiffUtils.diff(from.diffableRows(), to.diffableRows());
         List<Delta> deltas = patch.getDeltas();
         if (!deltas.isEmpty()) {
             Map<Integer, Delta> deltasByLine = createDeltasByLine(deltas);
-            throw new TableDiffException(createTableDiff(deltasByLine));
+            throw new TableDiffException(from, to, createTableDiff(deltasByLine));
         }
     }
 
@@ -48,18 +48,18 @@ public class TableDiffer {
 
     private DataTable createTableDiff(Map<Integer, Delta> deltasByLine) {
         List<DataTableRow> diffTableRows = new ArrayList<DataTableRow>();
-        List<List<String>> rows = orig.raw();
+        List<List<String>> rows = from.raw();
         for (int i = 0; i < rows.size(); i++) {
             Delta delta = deltasByLine.get(i);
             if (delta == null) {
-                diffTableRows.add(orig.getGherkinRows().get(i));
+                diffTableRows.add(from.getGherkinRows().get(i));
             } else {
                 addRowsToTableDiff(diffTableRows, delta);
                 // skipping lines involved in a delta
                 if (delta.getType() == Delta.TYPE.CHANGE || delta.getType() == Delta.TYPE.DELETE) {
                     i += delta.getOriginal().getLines().size() - 1;
                 } else {
-                    diffTableRows.add(orig.getGherkinRows().get(i));
+                    diffTableRows.add(from.getGherkinRows().get(i));
                 }
             }
         }
@@ -68,7 +68,7 @@ public class TableDiffer {
         if (remainingDelta != null) {
             addRowsToTableDiff(diffTableRows, remainingDelta);
         }
-        return new DataTable(diffTableRows, orig.getTableConverter());
+        return new DataTable(diffTableRows, from.getTableConverter());
     }
 
     private void addRowsToTableDiff(List<DataTableRow> diffTableRows, Delta delta) {